Move location API tests to a separate functional test.
The API requests are incompatible with the javascript test driver.
